Introduced the master and import scripts tailored for computing and  plotting traits of the process over many directories
The proliferation_magnitude parameter is now used in two ways.

If division is exponential,(misleading)
it sets conversion_t: determining the speed of the MF. The proliferation is changed to account for the rescaling of t_mech but
in such a manner so thaht it is the same as the experiment.

If division is uniform or no setting, then it changes the growth rate while keeping the MF speed the same.

The non_realistic param is on

The dir analysis scripts :

units are rescaled properly only on cross_dir not on indiv dir plots

MF speed test is from the no setting in proliferation and does not need rescaling with proliferation magnitude